ft(comment): create add comment route
