+++ src/__mocks__/comment.js
@@ -1,14 +1,26 @@
-export const comment = {
-    valid: {
-        message: 'Test message for comments',
-        quotedCommentId: null,
-        createdAt: new Date(),
-        updatedAt: new Date()
-    },
-    noMessage: {
-        message: '',
-        quotedCommentId: null,
-        createdAt: new Date(),
-        updatedAt: new Date()
-    }
-}
+export default {
+  valid: {
+    message: 'Test message for comments',
+    quotedCommentId: null,
+    createdAt: new Date(),
+    updatedAt: new Date()
+  },
+  quotedComment: {
+    message: 'Quoted Comment',
+    quotedCommentId: 1,
+    createdAt: new Date(),
+    updatedAt: new Date()
+  },
+  noMessage: {
+    message: '',
+    quotedCommentId: null,
+    createdAt: new Date(),
+    updatedAt: new Date()
+  },
+  invalidQuote: {
+    message: 'Test comment',
+    quotedCommentId: 'h',
+    createdAt: new Date(),
+    updatedAt: new Date()
+  }
+};
